Modificación del componente de creación de APU's ProjectCreate

+++ database/migrations/2024_03_05_220752_create_position_project_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('position_id');
             $table->unsignedBigInteger('project_id');
-            $table->integer('quantity')->default(0)->nullable(false);
-            $table->integer('required_days')->default(0)->nullable(false);
+            $table->bigInteger('quantity')->default(0)->nullable(false);
+            $table->bigInteger('required_days')->default(0)->nullable(false);
             $table->decimal('efficiency', 8, 2)->default(1.0)->nullable(false);
             $table->decimal('total_cost', 20, 2)->default(0)->nullable(false);
             $table->timestamps();
