+++ database/migrations/2024_03_05_223203_create_material_project_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('project_id');
             $table->unsignedBigInteger('material_id');
-            $table->integer('quantity')->default(0)->nullable(false);
+            $table->bigInteger('quantity')->default(0)->nullable(false);
             $table->decimal('efficiency', 8, 2)->default(1.0)->nullable(false);
             $table->decimal('total_cost', 20, 2)->default(0)->nullable(false);
             $table->timestamps();
